Process improvement for speed
I was transmitting the entire face image from the server to the client.
For large faces (~2kx2k pixels), this was taking tens of seconds. I
realized that I just have to adjust face_rect to have the bounds for
these boxes and methods to reconstruct them on the client side, which is
a huge bandwidth and pickling savings.